This implements `-r` and `--require` as command line options
@Raynos accurately pointed out that the source map support was a
bit too specific, and felt outside of the purview of tape.
As a more general solution, this implements the `-r` and `--require`
CLI flags, which will allow a user to load node modules before any
tests are run. For example, if someon wants to enable source map
support in node, they might do:

    tape -r source-map-support/register test/*.js