[New] `bin/tape`: add `--ignore-pattern` flag
This PR introduces the --ignore-pattern flag as a shorthand for ignoring test files.

The flag may be used together with --ignore; the input will be concatenated in that case. I figured this behavior would make the most sense, since users may want to ignore everything in .gitignore and some other files on top.

Fixes #586